pyproject.toml
[project]
name = "aak-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "uritemplate>=4.1.1",
]
